requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
requests

# Google Calendar API dependencies
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
